javacryptotools v0.1.0 - Add Counterparty (XCP) support & interfaces refactor
* Add XCP explorer Xchain.io
* Explorer, NftExplorer & PriceExplorer becomes interface instead of abstract classes
* get_token_balance(addr, contract) returns double instead of long

+++ javacryptotools/src/main/java/org/satochip/javacryptotools/coins/BaseCoin.java
@@ -217,11 +217,11 @@ public abstract class BaseCoin {
         }
     }
     
-    public long getTokenBalance(String addr, String contract){
+    public double getTokenBalance(String addr, String contract){
         if (explorer!=null){ 
             return explorer.get_token_balance(addr, contract);
         } else {
-            return (long)-1;
+            return (double)-1;
         }
     }
